Compile all engine modules under CMake/clang (macOS arm64)
Every module now compiles cross-platform: Units, Environment, Water,
UserInterface, 3dx, Render and Game (the build now reaches the final
exe link, which still needs stub/backend symbols).

Key changes:
- Expand fake d3d9.h/d3dx9.h so the D3D render headers parse everywhere
  (full IDirect3DDevice9 method set, COM base, D3DFMT_/D3DRS_/D3DTSS_
  constants, D3DX stubs) -- the real D3D backend stays WIN32-gated.
- WindowsAPI.h: large Win32 shim expansion (window/registry/COM/message
  -loop/cursor APIs, WM_*/VK_*/WS_* constants, lstr*/swprintf/_wtoi).
- New fake headers: shlobj, shellapi, ddraw, malloc (+ existing shlwapi,
  vfw); add -Wno-error=address-of-temporary (pervasive MSVC-ism in UI).
- Port MMX/SSE FogOfWar to scalar; UnkObj refcount to __atomic_*; SEH
  __try and __asm int 3 to portable equivalents.
- Stub/gate Win32-only hosts: Water module (WaterStub.cpp), Render
  Runtime3D/stack walkers, FileImage AVI codec; exclude orphans
  (Line3d, ShadowVolume, particle.cpp -- not in Render.vcproj).
- Convert FT_Font.cpp to UTF-8 (wide cyrillic literals are a hard clang
  error); link vendored FreeType include dir into Render.
- Many clang-vs-MSVC fixes: removed-STL functors -> lambdas, round()
  -> int() casts, enum ': int' fwd-decls, typename/this-> two-phase
  lookup (sigslot), const& temporaries, missing/malformed includes.

+++ Render/src/Grass.cpp
@@ -369,9 +369,9 @@ void GrassMap::SetGrass(float x, float y, float radius, int texture, int count, 
 	if (!textureMap_)
 		return;
 	xassert(count>0 && count <=16);
-	int xx = round(x)>>grassMapShift;
-	int yy = round(y)>>grassMapShift;
-	int rad = round(radius)>>grassMapShift;
+	int xx = int(round(x))>>grassMapShift;
+	int yy = int(round(y))>>grassMapShift;
+	int rad = int(round(radius))>>grassMapShift;
 	int pix = (texture+1)<<textureCountShift;
 	Vect2i min(xx-rad,yy-rad);
 	Vect2i max(xx+rad,yy+rad);
@@ -681,10 +681,10 @@ void GrassMap::DownGrass(Vect3f pos, float radius)
 {
 	Vect2i tmin;
 	Vect2i tmax;
-	tmin.x = round(pos.x - radius)>>tileShift;
-	tmin.y = round(pos.y - radius)>>tileShift;
-	tmax.x = round(pos.x + radius)>>tileShift;
-	tmax.y = round(pos.y + radius)>>tileShift;
+	tmin.x = int(round(pos.x - radius))>>tileShift;
+	tmin.y = int(round(pos.y - radius))>>tileShift;
+	tmax.x = int(round(pos.x + radius))>>tileShift;
+	tmax.y = int(round(pos.y + radius))>>tileShift;
 	tmin.x = max(0,tmin.x);
 	tmin.y = max(0,tmin.y);
 	tmax.x = min(tileNumber_.x-1,tmax.x);
@@ -715,10 +715,10 @@ void GrassMap::DeleteGrass(Vect3f pos, float radius)
 {
 	Vect2i tmin;
 	Vect2i tmax;
-	tmin.x = round(pos.x - radius)>>tileShift;
-	tmin.y = round(pos.y - radius)>>tileShift;
-	tmax.x = round(pos.x + radius)>>tileShift;
-	tmax.y = round(pos.y + radius)>>tileShift;
+	tmin.x = int(round(pos.x - radius))>>tileShift;
+	tmin.y = int(round(pos.y - radius))>>tileShift;
+	tmax.x = int(round(pos.x + radius))>>tileShift;
+	tmax.y = int(round(pos.y + radius))>>tileShift;
 	tmin.x = max(0,tmin.x);
 	tmin.y = max(0,tmin.y);
 	tmax.x = min(tileNumber_.x-1,tmax.x);
